third_party/blink/web_tests/external/wpt/speculation-rules/prefetch/tentative/cookie-indices.https_include=unchanged-expected.txt

This is a testharness.js-based test.
[FAIL] Cookie-Indices should not prevent a prefetch from succeeding if the cookie has not changed.
  assert_equals: expected (string) "hello" but got (undefined) undefined
Harness: the test ran to completion.
